+++ packages/emoji/src/components/Emoji/JoyPixelEmojiImage.tsx
@@ -6,13 +6,13 @@ export default function JoyPixelEmojiImage({
   emoji,
   theme,
 }: EmojiImageProps): ReactElement {
-  const imgTag = toImage(emoji);
+  const imgTag = toImage(emoji.shortname);
   const path = /src="(.*)"/.exec(imgTag)?.[1];
   return (
     <img
       src={path}
       className={theme.emojiSelectPopoverEntryIcon}
-      title={emoji}
+      title={emoji.shortname}
       draggable={false}
       role="presentation"
     />