+++ packages/emoji/src/components/EmojiSuggestions/index.tsx
@@ -27,6 +27,7 @@ import defaultPositionSuggestions, {
   PositionSuggestionsParams,
 } from '../../utils/positionSuggestions';
 import { EmojiPluginTheme } from '../../theme';
+import { EmojiShape } from '../../constants/type';
 import Popover from './Popover';
 import { warning } from '../../utils/warning';
 
@@ -42,7 +43,7 @@ interface EmojiSuggestionsParams extends EmojiSuggestionsPubParams {
   callbacks: EmojiPLuginCallbacks;
   ariaProps: AriaProps;
   store: EmojiPluginStore;
-  shortNames: List<string>;
+  emojis: List<EmojiShape>;
   positionSuggestions?(arg: PositionSuggestionsParams): CSSProperties;
   theme: EmojiPluginTheme;
   emojiImage: ComponentType<EmojiImageProps>;
@@ -57,7 +58,7 @@ export default class EmojiSuggestions extends Component<EmojiSuggestionsParams> 
 
   popover?: HTMLDivElement | null;
   key!: string;
-  filteredEmojis?: List<string>;
+  filteredEmojis?: List<EmojiShape>;
   activeOffsetKey?: string;
   lastSelectionIsInsideWord?: Immutable.Iterable<string, boolean>;
   lastSearchValue?: string;
@@ -93,7 +94,9 @@ export default class EmojiSuggestions extends Component<EmojiSuggestionsParams> 
           decoratorRect,
           props: this.props,
           state: this.state,
-          filteredEmojis: this.filteredEmojis,
+          filteredEmojis: this.filteredEmojis
+            .map((emojiShape) => emojiShape!.shortname)
+            .toList(),
           popover: this.popover,
         });
         for (const [key, value] of Object.entries(newStyles)) {
@@ -260,7 +263,7 @@ export default class EmojiSuggestions extends Component<EmojiSuggestionsParams> 
     this.props.store.setEditorState!(this.props.store.getEditorState!());
   };
 
-  onEmojiSelect = (emoji: string): void => {
+  onEmojiSelect = (emoji: EmojiShape): void => {
     this.closeDropdown();
     const newEditorState = addEmoji(
       this.props.store.getEditorState!(),
@@ -280,17 +283,17 @@ export default class EmojiSuggestions extends Component<EmojiSuggestionsParams> 
   };
 
   // Get the first 6 emojis that match
-  getEmojisForFilter = (): List<string> => {
+  getEmojisForFilter = (): List<EmojiShape> => {
     const selection = this.props.store.getEditorState!().getSelection();
     const { word } = getSearchText(
       this.props.store.getEditorState!(),
       selection
     );
     const emojiValue = word.substring(1, word.length).toLowerCase();
-    const filteredValues = this.props.shortNames.filter(
-      (emojiShortName) =>
-        !emojiValue || emojiShortName!.indexOf(emojiValue) > -1
-    ) as List<string>;
+    const filteredValues = this.props.emojis.filter(
+      (emojiShape) =>
+        !emojiValue || emojiShape!.shortname.indexOf(emojiValue) > -1
+    ) as List<EmojiShape>;
     const size = filteredValues.size < 9 ? filteredValues.size : 9;
     return filteredValues.setSize(size);
   };
@@ -372,7 +375,7 @@ export default class EmojiSuggestions extends Component<EmojiSuggestionsParams> 
       onSearchChange, // eslint-disable-line @typescript-eslint/no-unused-vars
       positionSuggestions, // eslint-disable-line @typescript-eslint/no-unused-vars
       popperOptions, // eslint-disable-line @typescript-eslint/no-unused-vars
-      shortNames, // eslint-disable-line @typescript-eslint/no-unused-vars
+      emojis, // eslint-disable-line @typescript-eslint/no-unused-vars
       store, // eslint-disable-line @typescript-eslint/no-unused-vars
       emojiImage,
       ...restProps
@@ -396,7 +399,7 @@ export default class EmojiSuggestions extends Component<EmojiSuggestionsParams> 
           {this.filteredEmojis
             .map((emoji, index) => (
               <Entry
-                key={emoji}
+                key={emoji!.shortname}
                 onEmojiSelect={this.onEmojiSelect}
                 onEmojiFocus={this.onEmojiFocus}
                 isFocused={this.state.focusedOptionIndex === index}
@@ -421,7 +424,7 @@ export default class EmojiSuggestions extends Component<EmojiSuggestionsParams> 
         {this.filteredEmojis
           .map((emoji, index) => (
             <Entry
-              key={emoji}
+              key={emoji!.shortname}
               onEmojiSelect={this.onEmojiSelect}
               onEmojiFocus={this.onEmojiFocus}
               isFocused={this.state.focusedOptionIndex === index}