+++ packages/emoji/src/index.tsx
@@ -13,7 +13,8 @@ import {
   EditorCommand,
 } from '@draft-js-plugins/editor';
 import { Map, List } from 'immutable';
-import keys from 'lodash/keys';
+import flatten from 'lodash/flatten';
+import values from 'lodash/values';
 import { DraftHandleValue, EditorState } from 'draft-js';
 import Emoji, { EmojiProps } from './components/Emoji';
 import EmojiSuggestions, {
@@ -35,6 +36,7 @@ import JoyPixelEmojiImage from './components/Emoji/JoyPixelEmojiImage';
 import NativeEmojiImage from './components/Emoji/NativeEmojiImage';
 import NativeEmojiInlineText from './components/Emoji/NativEmojiInlineText';
 import JoyPixelEmojiInlineText from './components/Emoji/JoyPixelEmojiInlineText';
+import { EmojiShape } from './constants/type';
 
 export { defaultTheme };
 export type { EmojiPluginTheme };
@@ -45,9 +47,8 @@ export type PopperOptions = Omit<Partial<PopperJS.Options>, 'modifiers'> & {
 };
 
 export interface EmojiImageProps {
-  emoji: string;
+  emoji: EmojiShape;
   theme: EmojiPluginTheme;
-  unicode: string;
 }
 
 export interface EmojiInlineTextProps {
@@ -203,7 +204,7 @@ export default (config: EmojiPluginConfig = {}): EmojiPlugin => {
     theme,
     store,
     positionSuggestions,
-    shortNames: List(keys(emojiList.list)),
+    emojis: List(flatten(values(emojiList.list))),
     emojiImage,
     popperOptions,
   };