+++ packages/emoji/src/utils/emojiList.ts
@@ -1,16 +1,20 @@
 import emojiToolkit from 'emoji-toolkit';
+import { EmojiShape } from 'packages/emoji/src/constants/type';
 
 interface EmojiListObject {
+  [s: string]: EmojiShape[];
+}
+interface EmojiPriorityListObject {
   [s: string]: string[];
 }
 
 interface EmojiList {
-  setPriorityList(newPriorityList: EmojiListObject): void;
+  setPriorityList(newPriorityList: EmojiPriorityListObject): void;
   list: EmojiListObject;
 }
 
 function newEmojiListWithOutPriorityList(
-  priorityList: EmojiListObject
+  priorityList: EmojiPriorityListObject
 ): EmojiListObject {
   const list: EmojiListObject = {};
   for (const key in emojiToolkit.emojiList) {
@@ -20,10 +24,26 @@ function newEmojiListWithOutPriorityList(
       continue; // eslint-disable-line no-continue
     }
 
-    list[key] = [emojiToolkit.emojiList[key].uc_base];
+    list[key] = [
+      {
+        shortname: key,
+        unicode: emojiToolkit.convert(emojiToolkit.emojiList[key].uc_full),
+      },
+    ];
   }
 
-  return { ...priorityList, ...list };
+  const mappedPriorityList = Object.entries(priorityList).reduce(
+    (acc, [shortname, unicodes]) => ({
+      ...acc,
+      [shortname]: unicodes.map((unicode) => ({
+        shortname,
+        unicode: emojiToolkit.convert(unicode),
+      })),
+    }),
+    {}
+  );
+
+  return { ...mappedPriorityList, ...list };
 }
 
 const emojiList: EmojiList = {
@@ -34,10 +54,10 @@ const emojiList: EmojiList = {
 };
 
 // init emojiList
-const priorityList: EmojiListObject = {
+const priorityList: EmojiPriorityListObject = {
   ':thumbsup:': ['1f44d'],
   ':smile:': ['1f604'],
-  ':heart:': ['2764-fe0f', '2764'],
+  ':heart:': ['2764-fe0f'],
   ':ok_hand:': ['1f44c'],
   ':joy:': ['1f602'],
   ':tada:': ['1f389'],